pyproject.toml
[tool.poetry]
name = "il-elections"
version = "0.1.0"
description = "Analyzing Israel elections data."
authors = ["Zach Moshe <zachmoshe@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.21.1"
pandas = "^1.3.2"
xlrd = "^2.0.1"
openpyxl = "^3.0.7"
parametrized = "^0.1"
parameterized = "^0.8.1"
absl-py = "^0.13.0"
googlemaps = "^4.5.3"
importlib_resources = "^5.2.2"
pyarrow = "^5.0.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
ipykernel = "^6.2.0"
nbconvert = "^6.1.0"
jupyter_contrib_nbextensions = "^0.5.1"
parameterized = "^0.8.1"
# Colab requirements
jupyter_http_over_ws = "^0.0.8"  # also requires running `jupyter serverextension enable --py jupyter_http_over_ws`
matplotlib = "^3.4.3"
pytest = "^6.2.5"
pylint = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
